Refactor local cloud account repository
